+++ python/__init__.py
@@ -1,10 +1,10 @@
-# Copyright 2023 The Bazel Authors. All rights reserved.
+# Copyright 2024 The Bazel Authors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#    http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,5 +13,11 @@
 # limitations under the License.
 
 import sys
-# todo: i think this needs to be special cased for workspace builds
-sys.modules["python.runfiles"] = rules_python.python.runfiles
+
+# todo: I think, for workspace builds, we need do something like:
+# if name == rules_python.python
+#   import from src.d/rules_python somehow
+# todo: i think we need to skip this for workspace builds?
+import rules_python.python
+
+sys.modules["python"] = rules_python.python
