+++ src.d/__init__.py
@@ -1,0 +1,4 @@
+raise ImportError(
+    "@rules_python//src.d:__init__.py should not be imported or importable. "
+    'You likely meant "import rules_python"'
+)
