+++ src.d/rules_python/__init__.py
@@ -1,10 +1,10 @@
-# Copyright 2023 The Bazel Authors. All rights reserved.
+# Copyright 2024 The Bazel Authors. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#    http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,6 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import sys
-# todo: i think this needs to be special cased for workspace builds
-sys.modules["python.runfiles"] = rules_python.python.runfiles
+# Intentionally empty.
+# rules_python should be treated as a namespace package, though it
+# doesn't quite function as one in practice.
